+++ src/lib/neo/convolution/dense_fdl.hpp
@@ -48,8 +48,8 @@ struct dense_split_fdl
         auto real = stdex::submdspan(_fdl.to_mdspan(), 0, index, stdex::full_extent);
         auto imag = stdex::submdspan(_fdl.to_mdspan(), 1, index, stdex::full_extent);
         for (auto i{0}; i < static_cast<int>(input.extent(0)); ++i) {
-            real[i] = input[i].real();
-            imag[i] = input[i].imag();
+            real[i] = static_cast<Float>(input[i].real());
+            imag[i] = static_cast<Float>(input[i].imag());
         }
     }
